+++ src/GRA.Domain.Model/ImportStatus.cs
@@ -1,0 +1,10 @@
+﻿namespace GRA.Domain.Model
+{
+    public enum ImportStatus
+    {
+        Success,
+        Info,
+        Warning,
+        Danger
+    }
+}