+++ src/GRA.Domain.Model/SchoolImportExport.cs
@@ -1,0 +1,9 @@
+﻿namespace GRA.Domain.Model
+{
+    public class SchoolImportExport
+    {
+        public string Type { get; set; }
+        public string District { get; set; }
+        public string Name { get; set; }
+    }
+}